Fix a bug with min-utxo outputs
For convenience, we want to allow payments to be constructed without worrying about the minUTXO; if the calculated amount falls below the minUTXO, we increase it to the minimum, and let tx balancing find the ada to cover it.  However, in the current implementation, we were coercing the minutxo amount when calculating the `requestedAmount`, in .ToValue.  But, we weren't doing the same thing in .ToTxOut(), which would leave the tx imbalanced.  To solve this, I've created a dedicated EnsureMinUTXO method, which ensures that after it gets massaged, it'll be consistent across all future uses